refactor: changed dataset.py to prep.py - include load_data and build_features
